finish CompilationCommit

+++ main-server/src/main/java/ru/practicum/user/service/UserServiceImpl.java
@@ -78,6 +78,9 @@ public class UserServiceImpl implements UserService {
     @Override
     @Transactional
     public void delete(Long id) {
+        if (!userRepository.existsById(id)) {
+            throw new NotFoundException("Не найден пользователь с id: " + id);
+        }
         log.info("Удаление пользователя с id: {}", id);
         userRepository.deleteById(id);
     }
